feat: integrate app with db
