+++ cart-service/package.json
@@ -13,7 +13,7 @@
     "start:dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/main",
-    "offline": "nest build && sls offline",
+    "offline": "npm run build && sls offline",
     "deploy": "nest build && sls deploy --aws-profile main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "test": "jest",
@@ -31,20 +31,24 @@
     "@nestjs/platform-express": "^7.5.1",
     "@vendia/serverless-express": "^4.10.1",
     "aws-lambda": "^1.0.7",
+    "bcryptjs": "^2.4.3",
     "helmet": "^4.2.0",
     "passport": "^0.4.1",
     "passport-http": "^0.3.0",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
+    "pg": "^8.8.0",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^6.5.4",
-    "uuid": "^8.3.1"
+    "uuid": "^8.3.1",
+    "validator": "^13.7.0"
   },
   "devDependencies": {
     "@nestjs/cli": "^7.5.2",
     "@nestjs/schematics": "^7.2.1",
     "@nestjs/testing": "^7.5.1",
     "@types/aws-lambda": "^8.10.108",
+    "@types/bcryptjs": "^2.4.2",
     "@types/express": "^4.17.3",
     "@types/jest": "26.0.10",
     "@types/node": "^13.13.30",
