+++ cart-service/src/shared/db/index.ts
@@ -1,0 +1,11 @@
+import { Pool } from 'pg';
+
+const { PG_HOST, PG_PORT, PG_DATABASE, PG_USER, PG_PASSWORD } = process.env;
+
+export const dbPool = new Pool({
+  user: PG_USER,
+  host: PG_HOST,
+  database: PG_DATABASE,
+  password: PG_PASSWORD,
+  port: PG_PORT,
+});
